reorg SDK codebase for clarity
