+++ packages/zkpassport-sdk/src/constants.ts
@@ -1,0 +1,3 @@
+export const VERSION = "0.9.0"
+export const DEFAULT_VALIDITY = 7 * 24 * 60 * 60 // 7 days
+export const DEFAULT_DATE_VALUE = new Date(0)
