+++ packages/zkpassport-utils/src/types/index.ts
@@ -332,6 +332,7 @@ export type CommittedInputs =
   | DiscloseCommittedInputs
   | BindCommittedInputs
   | SanctionsCommittedInputs
+  | FacematchCommittedInputs
 
 export type DisclosureCircuitName =
   | "disclose_bytes"
